refactor: header와 footer를 app/layouts로 이동
